nixd/fork: fix regression diagnostic
