+++ lib/nixd/meson.build
@@ -4,7 +4,9 @@ nixd_server_lib = library('nixd-lsp'
   , 'src/CallbackExpr.cpp'
   , 'src/Diagnostic.cpp'
   , 'src/EvalDraftStore.cpp'
-  , 'src/Server.cpp'
+  , 'src/ServerController.cpp'
+  , 'src/ServerSerialization.cpp'
+  , 'src/ServerWorker.cpp'
   ]
 , include_directories: nixd_server_inc
 , dependencies: [ llvm
